feat: Page de profil utilisateur

+++ delete-article.php
@@ -19,7 +19,11 @@
 
         
         if($idArticle) {
+         $article = $articleDAO->getOne($idArticle);
+         if($article['author'] === $currentUser['id']){
             $articleDAO->deleteOne($idArticle);
+
+         }
         }
 
         header('Location: /');
